Fix overflow error when parsing floats
When parsing floats with anything but an all-zero fractional part,
multiplying by 100_000_000 would panic with an overflow error due to the
fractional part of a Real only being a u32. Converting the fractional
part to be a u64 instead solves this.

+++ noria/noria-logictest/src/parser.rs
@@ -165,7 +165,7 @@ named!(
         whole: flat_map!(digit1, parse_to!(i64))
             >> tag!(".")
             >> fractional: flat_map!(digit1, parse_to!(u32))
-            >> (Value::Real(whole, fractional * 100_000_000))
+            >> (Value::Real(whole, (fractional as u64) * 100_000_000))
     )
 );
